# swagger/docs.json
{
  "openapi": "3.0.2",
  "info": {
    "title": "CineReviews",
    "version": "1.0"
  },
  "servers": [
    {
      "url": "http://localhost:3000/",
      "description": "Local Server"
    }
  ],
  "paths": {
    "/movies": {
      "get": {
        "description": "Get movies",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "$ref": "#/components/schemas/page"
          },
          {
            "name": "size",
            "in": "query",
            "required": false,
            "$ref": "#/components/schemas/size"
          },
          {
            "name": "sortParam",
            "in": "query",
            "required": false,
            "$ref": "#/components/schemas/sortParam"
          },
          {
            "name": "sortDirection",
            "in": "query",
            "required": false,
            "$ref": "#/components/schemas/sortDirection"
          }
        ],
        "responses": {
          "200": {
            "description": "Movies were obtained",
            "content": {
              "text/html": {
                "schema": {
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "text/html": {
                "schema": {
                }
              }
            }
          }
        }
      }
    },
    "/movies/name": {
      "get": {
        "description": "Get movies by name",
        "parameters": [
          {
            "name": "partial",
            "in": "query",
            "required": true,
            "$ref": "#/components/schemas/partial"
          }
        ],
        "responses": {
          "200": {
            "description": "Movies were obtained",
            "content": {
              "text/html": {
                "schema": {
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "text/html": {
                "schema": {
                }
              }
            }
          }
        }
      }
    },
    "/movies/id": {
      "get": {
        "description": "Get movie by name",
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "required": true,
            "$ref": "#/components/schemas/Movie/properties/id"
          }
        ],
        "responses": {
          "200": {
            "description": "Movies were obtained",
            "content": {
              "text/html": {
                "schema": {
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "text/html": {
                "schema": {
                }
              }
            }
          }
        }
      }
    },
    "/movies/search": {
      "get": {
        "description": "Get movie search page",
        "responses": {
          "200": {
            "description": "Movie search were obtained",
            "content": {
              "text/html": {
                "schema": {
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "text/html": {
                "schema": {
                }
              }
            }
          }
        }
      }
    },
    "/movies/filter/table": {
      "get": {
        "description": "Get handlebars table component",
        "responses": {
          "200": {
            "description": "handlebars table componentwere obtained",
            "content": {
              "text/html": {
                "schema": {
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "text/html": {
                "schema": {
                }
              }
            }
          }
        }
      }
    },
    "/movies/filter": {
      "get": {
        "description": "Get filterd data",
        "parameters": [
          {
            "name": "country",
            "in": "query",
            "required": false,
            "$ref": "#/components/schemas/Country/properties/id"
          },
          {
            "name": "genre",
            "in": "query",
            "required": false,
            "$ref": "#/components/schemas/Genre/properties/id"
          },
          {
            "name": "language",
            "in": "query",
            "required": false,
            "$ref": "#/components/schemas/Language/properties/id"
          },
          {
            "name": "theme",
            "in": "query",
            "required": false,
            "$ref": "#/components/schemas/Theme/properties/id"
          },
          {
            "name": "date",
            "in": "query",
            "required": false,
            "$ref": "#/components/schemas/Movie/properties/date"
          }
        ],
        "responses": {
          "200": {
            "description": "Movies were obtained",
            "content": {
              "text/html": {
                "schema": {
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "text/html": {
                "schema": {
                }
              }
            }
          }
        }
      }
    },
    "/movies/genres": {
      "get": {
        "description": "Get movies filtered by genre",
        "parameters": [
          {
            "name": "genreId",
            "in": "query",
            "required": false,
            "$ref": "#/components/schemas/Genre/properties/id"
          }
        ],
        "responses": {
          "200": {
            "description": "Movies were obtained",
            "content": {
              "text/html": {
                "schema": {
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "text/html": {
                "schema": {
                }
              }
            }
          }
        }
      }
    },
    "/movies/top10-movies": {
      "get": {
        "description": "Get top 10 movies based on parameters",
        "parameters": [
          {
            "name": "type",
            "in": "query",
            "required": true
          },
          {
            "name": "country",
            "in": "query",
            "required": false,
            "$ref": "#/components/schemas/Country/properties/id"
          },
          {
            "name": "genre",
            "in": "query",
            "required": false,
            "$ref": "#/components/schemas/Genre/properties/id"
          },
          {
            "name": "language",
            "in": "query",
            "required": false,
            "$ref": "#/components/schemas/Language/properties/id"
          }
        ],
        "responses": {
          "200": {
            "description": "Movies were obtained",
            "content": {
              "text/html": {
                "schema": {
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "text/html": {
                "schema": {
                }
              }
            }
          }
        }
      }
    },
...
    "/actors/movie": {
      "get": {
        "description": "Get actors for a certain movie",
        "parameters": [
          {
            "name": "movieId",
            "required": true,
            "in": "query",
            "$ref": "#/components/schemas/Movie/properties/id"
          }
        ],
        "responses": {
          "200": {
            "description": "Actors were obtained",
            "content": {
              "text/html": {
                "schema": {
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "text/html": {
                "schema": {
                }
              }
            }
          }
        }
      }
    },
...
    "/genres": {
      "get": {
        "description": "Retrieve a list of all genres.",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 0
            },
            "description": "The page number for pagination."
          },
          {
            "name": "size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 10
            },
            "description": "The number of items per page."
          }
        ],
        "responses": {
          "200": {
            "description": "A list of genres was retrieved successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Genre"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "integer",
                      "example": 500
                    },
                    "message": {
                      "type": "string",
                      "example": "Internal Server Error"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/genres/movie": {
      "get": {
        "description": "Retrieve genres for a specific movie by its ID.",
        "parameters": [
          {
            "name": "movieId",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "The ID of the movie."
          }
        ],
        "responses": {
          "200": {
            "description": "Genres for the specified movie were retrieved successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Genre"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "integer",
                      "example": 500
                    },
                    "message": {
                      "type": "string",
                      "example": "Internal Server Error"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/languages": {
      "get": {
        "description": "Retrieve a list of all languages.",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 0
            },
            "description": "The page number for pagination."
          },
          {
            "name": "size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 10
            },
            "description": "The number of items per page."
          }
        ],
        "responses": {
          "200": {
            "description": "A list of languages was retrieved successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Language"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "integer",
                      "example": 500
                    },
                    "message": {
                      "type": "string",
                      "example": "Internal Server Error"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "page": {
        "type": "integer",
        "description": "number of the page if Pagination is used"
      },
      "size": {
        "type": "integer",
        "description": "size of the page if Pagination is used"
      },
      "sortParam": {
        "type": "string",
        "description": "parameter used to filter the data"
      },
      "sortDirection": {
        "type": "string",
        "enum": [
          "ASC",
          "DESC"
        ],
        "description": "parameter used to chose the filtering order"
      },
      "partial": {
        "type": "string",
        "description": "part of the name of the movie"
      },
...
      "Genre": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "description": "Unique identifier for the genre",
            "example": 1
          },
          "genre": {
            "type": "string",
            "description": "Genre name",
            "example": "Comedy"
          }
        }
      },
...
      "Language": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "description": "Unique identifier for the language",
            "example": 1
          },
          "theme": {
            "type": "string",
            "description": "language specification",
            "example": "English"
          }
        }
      }
    }
  }
}